5380: rename and css import
