+++ base.scss
@@ -46,6 +46,8 @@
 @import "./src/stories/Library/Modals/modal-text/modal-text";
 @import "./src/stories/Library/Modals/modal-pause/modal-pause";
 @import "./src/stories/Library/material-page/material-page";
+@import "./src/stories/Library/toggle-button/toggle-button";
+@import "./src/stories/Library/pause-reservation/pause-reservation";
 @import "./src/stories/Library/material-header/material-header";
 @import "./src/stories/Library/material-header/material-periodikum-select";
 @import "./src/stories/Library/material-description/material-description";
